Add light/dark mode switch button (#22)
If the target mode is same as current system mode the new mode is the system mode, so further change to the system light/dark mode will change the app light/dark mode. A color filter is used to change the colors of the map in dark mode.

Also apply AppBar color to the theme to make it same as the old behavior.